Update party per person using pdpy
